Subscribers Model

+++ db/migrate/20170430144150_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -1,10 +1,10 @@
 # This migration comes from acts_as_taggable_on_engine (originally 2)
-class AddMissingUniqueIndices < ActiveRecord::Migration
+class AddMissingUniqueIndices < ActiveRecord::Migration[5.2]
   def self.up
     add_index :tags, :name, unique: true
 
     remove_index :taggings, :tag_id if index_exists?(:taggings, :tag_id)
-    remove_index :taggings, [:taggable_id, :taggable_type, :context]
+    remove_index :taggings, [:taggable_id, :taggable_type, :context] if index_exists?(:taggings, [:taggable_id, :taggable_type, :context])
     add_index :taggings,
               [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type],
               unique: true, name: 'taggings_idx'
